Fix slider value reassignment bug
The issue was the the _update_slider method is handled on resizing. binding a value to a function does not store a reference to a variable but rather clones it at "bind-time". This means that the _update_slider function bound to "resized" just permanently set itself back to the default value because it would be reading from the value set at bind-time.

I replaced this with a lambda function so that the actual value in "value" is respected at signal emission time, not bind-time.