Fix slider clipping bug
This is a little hacky but I tried to tweak the sliders as little as possible. The "Outerline" node has been replaced with a Polygon2D that is placed behind the ProgressBar in the SceneTree which is being moved by a RemoteTransform2D attached to the grabber.

This is the first solution that jumped out at me, and it *does* work, but idk if the approach is too unorthodox for this project