Action Rebinding, The Rebind:tm:
a commit ive been putting off for a long, long time. the reason i redesigned all of the pause menu UI in the first place: action rebinds. this commit is super super complicated and convoluted and UNFINISHED. since its incredibly ambitious.

for now, saving and loading bindings doesn't work yet, and only Keyboard Keys are supported (not yet MouseButtons like the UI says)

some functionality for saving and reloading is there but i need sleep before i continue with any of that

- Added a new class: AnimatedTextureRect, allows you to use SpriteFrames for the texture of a TextureRect.
- UIBindButton and BindSetting, the former of which handles generalised bind input to attach to InputMap actions.